Move the tree compression code to a separate file
